Feature/rec 29/create recipe

+++ RecipeApi/Service/RecipeService.cs
@@ -1,9 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using RecipeApi.Authentication;
 using RecipeApi.Database;
 using RecipeApi.Exceptions;
 using RecipeApi.IService;
@@ -16,20 +21,28 @@ namespace RecipeApi.Service
         private readonly RecipeDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
-        public RecipeService(RecipeDbContext context, IMapper mapper, ILogger<RecipeService> logger)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly UserManager<ApplicationUser> _userManager;
+        public RecipeService(
+            RecipeDbContext context,
+            IMapper mapper,
+            ILogger<RecipeService> logger,
+            IAuthorizationService authorizationService,
+            UserManager<ApplicationUser> userManager)
         {
             _dbContext = context;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userManager = userManager;
         }
-        public int AddRecipe(CreateRecipeDto recipe)
+        public int AddRecipe(CreateRecipeDto recipe, string userId)
         {
 
-            if(recipe is null) throw new BadRequestException("Empty recipe data");
+            if (recipe is null) throw new BadRequestException("Empty recipe data");
 
             var recipeEntity = _mapper.Map<Recipe>(recipe);
-            //TODO: temporary 
-            recipeEntity.UserId = 4;
+            recipeEntity.UserId = userId;
             recipeEntity.CreatedAt = DateTime.Now.ToUniversalTime();
             _dbContext.Recipes.Add(recipeEntity);
             _dbContext.SaveChanges();
@@ -37,12 +50,20 @@ namespace RecipeApi.Service
             return recipeEntity.Id;
         }
 
-        public void DeleteRecipe(int id)
+        public void DeleteRecipe(int id, string userId)
         {
             _logger.LogInformation($"Deleting recipe with id {id}");
             var recipe = _dbContext.Recipes.FirstOrDefault(r => r.Id == id);
             if (recipe is null) throw new NotFoundException("Recipe not found");
+
+            if(userId!=recipe.UserId) throw new ForbidException("You are not authorized to perform this action");
+
+            var recipeIngredients = _dbContext.RecipeIngredients.Where(ri => ri.RecipeId == id);
+            var recipeInstructions = _dbContext.RecipeInstructions.Where(ri => ri.RecipeId == recipe.Id);
+            
             _dbContext.Recipes.Remove(recipe);
+            _dbContext.RecipeIngredients.RemoveRange(recipeIngredients);
+            _dbContext.RecipeInstructions.RemoveRange(recipeInstructions);
             _dbContext.SaveChanges();
         }
 
@@ -73,7 +94,7 @@ namespace RecipeApi.Service
             return _mapper.Map<IEnumerable<RecipeDto>>(recipes);
         }
 
-        public IEnumerable<RecipeDto> GetRecipesByAuthor(int authorId)
+        public IEnumerable<RecipeDto> GetRecipesByAuthor(string authorId)
         {
             var recipes = _dbContext.Recipes.Where(r => r.UserId == authorId)
             .Include(r => r.User)
@@ -81,18 +102,23 @@ namespace RecipeApi.Service
             .Include(r => r.Instructions)
             .ToList();
 
-            if (recipes.Count==0) throw new NotFoundException("No recipes found for this author");
+            if (recipes.Count == 0) throw new NotFoundException("No recipes found for this author");
 
             return _mapper.Map<IEnumerable<RecipeDto>>(recipes);
         }
 
-        public void UpdateRecipe(int id, UpdateRecipeDto recipe)
+        public void UpdateRecipe(int id, UpdateRecipeDto recipe, string userId)
         {
             _logger.LogInformation($"Updating recipe with id {id}");
 
             if (recipe is null) throw new BadRequestException("Invalid recipe");
+
             var recipeEntity = _dbContext.Recipes.FirstOrDefault(r => r.Id == id);
             if (recipeEntity is null) throw new NotFoundException("Recipe not found");
+
+            if(userId!=recipeEntity.UserId) throw new ForbidException("You are not authorized to perform this action");
+            
+
             recipeEntity.Title = recipe.Title;
             recipeEntity.Description = recipe.Description;
             recipeEntity.Portions = recipe.Portions;
